src/main/java/com/univ/student/controller/StudentController.java
package com.univ.student.controller;

import com.univ.student.domain.model.StudentDto;
import com.univ.student.service.StudentService;
import lombok.RequiredArgsConstructor;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;
import java.util.List;

@RestController
@RequestMapping("/student")
@RequiredArgsConstructor
public class StudentController {

    private final StudentService studentService;

    @PostMapping
    public ResponseEntity<Object> createStudent(@Valid @RequestBody StudentDto studentDto) {
        StudentDto createdStudent = studentService.createStudent(studentDto);
        return ResponseEntity.ok(createdStudent);
    }

    @GetMapping
    public ResponseEntity<List<StudentDto>> getAllStudents() {
        return ResponseEntity.ok(studentService.getStudentList());
    }

    @GetMapping("/id/{id}")
    public ResponseEntity<StudentDto> getStudent(@PathVariable Long id) {
        return ResponseEntity.ok(studentService.getStudent(id));
    }

    @GetMapping("/departmentName/{departmentName}")
    public ResponseEntity<List<StudentDto>> getStudent(@PathVariable String departmentName) {
        return ResponseEntity.ok(studentService.getStudentByDepartmentName(departmentName));
    }
}
